move sql script from presentation to domain layer + infra
