+++ src/domain/repositories/CategoryOptionComboDeleteExporter.ts
@@ -1,0 +1,5 @@
+import { Id } from "domain/entities/Base";
+
+export interface CategoryOptionComboDeleteExporter {
+    exportDeleteScript(cocIds: Id[]): void;
+}
